ugc_service/requirements.txt
fastapi==0.111.0
pydantic-settings==2.10.1
gunicorn==23.0.0
uvicorn==0.38.0
pymongo==4.15.3
beanie==2.0.0
motor==3.7.1
sentry-sdk[fastapi]>=1.0.0
python-logstash==0.4.8
sentry-sdk==2.51.0
python-logstash-async
python-jose[cryptography]==3.5.0
pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1
asyncio==4.0.0